+++ packages/bluesky_cli/lib/src/commands/codegen/app/bsky/feed/send_interactions.dart
@@ -21,7 +21,7 @@ final class SendInteractionsCommand extends ProcedureCommand {
   SendInteractionsCommand() {
     argParser
       ..addOption("feed")
-      ..addMultiOption("interactions");
+      ..addMultiOption("interactions", splitCommas: false);
   }
 
   @override