+++ packages/bluesky_cli/lib/src/commands/codegen/chat/bsky/convo/send_message_batch.dart
@@ -19,7 +19,7 @@ import '../../../../procedure_command.dart';
 
 final class SendMessageBatchCommand extends ProcedureCommand {
   SendMessageBatchCommand() {
-    argParser..addMultiOption("items");
+    argParser..addMultiOption("items", splitCommas: false);
   }
 
   @override